event_handler: split between server & client handlers
