+++ src/core/event_handler.hpp
@@ -3,11 +3,6 @@
 #include "reactphysics3d/reactphysics3d.h"
 #include "registry/registry.hpp"
 
-struct RigidBodyData {
-    RigidBodyData(entt::entity aEntity) : Entity(aEntity) {}
-    entt::entity Entity;
-};
-
 class EventHandler : public rp3d::EventListener
 {
    public:
@@ -21,16 +16,4 @@ class EventHandler : public rp3d::EventListener
 
    private:
     Registry* mRegistry;
-};
-
-class TowerBuildingHandler : public rp3d::EventListener
-{
-   public:
-    /// Called when some contacts occur
-    /**
-     * @param callbackData Contains information about all the contacts
-     */
-    void onContact(const rp3d::CollisionCallback::CallbackData& /*callbackData*/) override;
-
-    bool CanBuildTower;
 };
